Update secret version and secret rotation to support using name_prefix in secret
Currently, secret version and secret rotation resources are set to each.key, which does not take into account the supported `name_prefix` parameter that can be set in the secrets' creation. 
Since the secret is already a dependency for these resources, we can use the output from the secret as the secret_id.